# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EmailGenie API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface EmailGenerationInput {
  intent: string;
  keyFacts: string;
  tone: string;
  businessName: string;
}

export interface EmailGenerationResult {
  modelA: string;
  modelB: string;
}

export interface EmailEvaluationInput {
  emailText: string;
  keyFacts: string;
  tone: string;
}

export interface EmailEvaluationResult {
  factRecallScore: number;
  toneAccuracyScore: number;
  professionalQualityScore: number;
  overallScore: number;
}

export type ShareInputData = { [key: string]: unknown };

export interface ShareInput {
  data: ShareInputData;
}

export interface ShareCreated {
  id: string;
}

export type ShareDataData = { [key: string]: unknown };

export interface ShareData {
  data: ShareDataData;
}