# lib/api-zod/src/generated/types/emailEvaluationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EmailGenie API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailEvaluationInput {
  emailText: string;
  keyFacts: string;
  tone: string;
}