# lib/api-zod/src/generated/types/emailEvaluationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EmailGenie API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailEvaluationResult {
  factRecallScore: number;
  toneAccuracyScore: number;
  professionalQualityScore: number;
  overallScore: number;
}