# lib/api-zod/src/generated/types/shareData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EmailGenie API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShareDataData } from './shareDataData';

export interface ShareData {
  data: ShareDataData;
}