# lib/api-zod/src/generated/types/shareInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EmailGenie API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShareInputData } from './shareInputData';

export interface ShareInput {
  data: ShareInputData;
}